lees ad in java + print armen/motor

+++ src/main/java/com/vdzon/robot/RobotUitlezing.java
@@ -3,5 +3,5 @@ package com.vdzon.robot;
 import com.vdzon.MoveRequest;
 
 public interface RobotUitlezing {
-    int getArmPos(int arm);
+    double getArmPos(int arm);
 }
